Move file to correct directory
